# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Finance API",
    "version": "1.0.0",
    "description": "API for categories, transactions, users and accounts"
  },
  "paths": {
...
    "/api/users": {
      "get": {
        "summary": "Get all users",
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "example": [
                  { "id": "1", "name": "Ivan" },
                  { "id": "2", "name": "Anna" }
                ]
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": { "name": "New User" }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" }
        }
      }
    },
...
    "/api/accounts": {
      "get": {
        "summary": "Get all accounts",
        "responses": {
          "200": {
            "description": "List of accounts",
            "content": {
              "application/json": {
                "example": [
                  { "id": "1", "name": "Main Account" },
                  { "id": "2", "name": "Savings" }
                ]
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": { "name": "New Account" }
            }
          }
        },
        "responses": {
          "201": { "description": "Account created" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}